+++ portal-ui/CHANGELOG.md
@@ -1,5 +1,7 @@
 # portal-ui
 
+## 0.1.1-beta.1
+
 ## 0.1.1-beta.0
 
 ### Patch Changes